requirements.txt
Django==5.0.1
django_allauth==0.60.1
python-dotenv==1.0.1
Requests==2.31.0
